dapps/b-cert/verifier/backend/requirements.txt
amqp
asgiref
astroid
billiard
cachetools
cairocffi
CairoSVG
celery
certifi
cffi
chardet
click
click-didyoumean
click-plugins
click-repl
cryptography
cssselect2
defusedxml
Django
django-allauth
django-cleanup
django-cors-headers
django-crispy-forms
django-model-utils
django-multiselectfield
django-template-check
djangorestframework
djangorestframework-jwt
djangorestframework-simplejwt
google-api-core
google-api-python-client
google-auth
google-auth-httplib2
googleapis-common-protos
httplib2
idna
importlib-metadata
install
isort
kombu
lazy-object-proxy
mccabe
oauth2client
oauthlib
pika
Pillow
progressbar2
prompt-toolkit
psycopg2
pyasn1
pyasn1-modules
pycparser
pydub
PyJWT
pylint
pylint-django
pylint-plugin-utils
python-dotenv
python-utils
python3-openid
pytz
PyYAML
redis
requests
requests-oauthlib
rsa
six
sqlparse
tinycss2
toml
tqdm
typed-ast
uritemplate
urllib3
vine
Wand
wcwidth
webencodings
wrapt
zipp
sawtooth_signing==1.1.5
bcrypt==3.1.7
pycryptodome==3.10.1
itsdangerous==1.1.0
protobuf==3.17.0
jsonschema==3.2.0
sawtooth_sdk==1.2.3
secp256k1==0.13.2
aiohttp
eciespy
pandas
numpy